+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Filter order/order lines by prefix or suffix. Refs UIOR-646.
 * Update stripes to v6.0.0 in Thunderjet's modules. Refs UIOR-650.
 * Add "Export results (CSV)" action to orders app - POL search. Refs UIOR-638.
 * Fix order not loading when system cannot get exchange rate for chosen currency. Refs UIOR-642